searcher/pyproject.toml
[tool.poetry]
name = "searcher"
version = "0.1.0"
description = ""
authors = ["Konstantin"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pika = "^1.3.2"
watchdog = "^4.0.0"
python-magic = "^0.4.27"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
